# payload/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Egregore Projection Plane",
    "version": "0.6.0-phase1"
  },
  "paths": {
...
    "/api/v1/ombudsman/route/{taxonomy}": {
      "get": {
        "tags": [
          "ombudsman"
        ],
        "summary": "Route To Cell",
        "description": "Find the best available cell for a given taxonomy path.",
        "operationId": "route_to_cell_api_v1_ombudsman_route__taxonomy__get",
        "parameters": [
          {
            "name": "taxonomy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taxonomy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Route To Cell Api V1 Ombudsman Route  Taxonomy  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/login": {
      "get": {
        "summary": "Login Page",
        "description": "Self-hosted login page: enter API key to receive a session cookie.",
        "operationId": "login_page_dashboard_login_get",
        "parameters": [
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Login Submit",
        "description": "Validate API key and set session cookie.",
        "operationId": "login_submit_dashboard_login_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_submit_dashboard_login_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/freeze": {
      "post": {
        "summary": "Freeze Action",
        "description": "Freeze the system.",
        "operationId": "freeze_action_dashboard_freeze_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_freeze_action_dashboard_freeze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/unfreeze": {
      "post": {
        "summary": "Unfreeze Action",
        "description": "Unfreeze the system.",
        "operationId": "unfreeze_action_dashboard_unfreeze_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_unfreeze_action_dashboard_unfreeze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/reset": {
      "post": {
        "summary": "Reset Action",
        "description": "Reset the system to HEALTHY.",
        "operationId": "reset_action_dashboard_reset_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_reset_action_dashboard_reset_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_freeze_action_dashboard_freeze_post": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "operator": {
            "type": "string",
            "title": "Operator",
            "default": "operator"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "Body_freeze_action_dashboard_freeze_post"
      },
      "Body_login_submit_dashboard_login_post": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "Body_login_submit_dashboard_login_post"
      },
      "Body_reset_action_dashboard_reset_post": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "operator": {
            "type": "string",
            "title": "Operator",
            "default": "operator"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "Body_reset_action_dashboard_reset_post"
      },
      "Body_unfreeze_action_dashboard_unfreeze_post": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "operator": {
            "type": "string",
            "title": "Operator",
            "default": "operator"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "Body_unfreeze_action_dashboard_unfreeze_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}